reorganize and cleanUp
