+++ idea-plugin/build.gradle
@@ -10,6 +10,7 @@ dependencies {
   implementation "org.jetbrains.kotlin:kotlin-stdlib:$kotlin_version"
   compile "org.celtric.kotlin:kotlin-html:0.1.4"
   compile project(path: ':compiler-plugin', configuration: 'shadow')
+  testImplementation project(":testing-plugin")
 }
 
 compileKotlin {
