+++ idea-plugin/src/test/kotlin/arrow/meta/ide/testing/dsl/IdeTestSyntax.kt
@@ -1,0 +1,6 @@
+package arrow.meta.ide.testing.dsl
+
+import arrow.meta.ide.testing.dsl.lineMarker.LineMarkerProviderTest
+import arrow.meta.ide.testing.env.IdeTestTypeSyntax
+
+interface IdeTestSyntax : IdeTestTypeSyntax, LineMarkerProviderTest
