--- idea-plugin/src/test/kotlin/arrow/meta/idea/test/code/comprehensions/IdeComprehensionsTest.kt
@@ -1,1 +1,0 @@
-package arrow.meta.idea.test.code.comprehensions
